Feature: IsMarried, occuptation details, age calc, new relations

+++ src/app/admin/contactus/contactus.component.ts
@@ -6,18 +6,19 @@ import { IContactusResult } from '../../model/contactus';
 @Component({
   selector: 'app-contactus',
   templateUrl: './contactus.component.html',
-  styleUrl: './contactus.component.css'
+  styleUrl: './contactus.component.css',
 })
 export class ContactusComponent implements OnInit {
-
   contacts$!: Observable<IContactusResult[]>;
-  constructor(private contactusService: ContactusService){}
+  constructor(private contactusService: ContactusService) {}
 
   ngOnInit(): void {
     this.contacts$ = this.contactusService.loadData();
   }
 
   onDelete(id: string) {
-    this.contactusService.deleteMessage(id);
+    if (confirm('Are you sure to delete?')) {
+      this.contactusService.deleteMessage(id);
+    }
   }
 }
